Refactor resumable hash support
To make this package easier to integrate in projects that would like to easily
choose between stdlib crypto and resumable digests, the package has been
refactored to simply register the resumable hashes with the main crypto
package. Users can then just access the hash functions through the standard
interface and type assert to detect support. The central package is now a
definition of a resumable hash to assist in this pattern.

Signed-off-by: Stephen J Day <stephen.day@docker.com>